fix: address review comment 4532002981 — split planning ops shard
